+++ tools/ci/tool-list.sh
@@ -120,14 +120,16 @@ case "$(uname -s)" in
   MINGW* | MSYS* | CYGWIN* | Windows_NT) host_os=windows ;;
   *) bail "unrecognized OS type '$(uname -s)'" ;;
 esac
-# See main.sh
+# NB: Sync with main.sh.
 case "$(uname -m)" in
   aarch64 | arm64) host_arch=aarch64 ;;
   xscale | arm | armv*l) bail "32-bit Arm runner is not supported yet by this action; if you need support for this platform, please submit an issue at <https://github.com/taiki-e/install-action>" ;;
-  *)
-    # TODO: uname -m on windows-11-arm returns "x86_64"
-    host_arch=x86_64
-    ;;
+  ppc64le) host_arch=powerpc64le ;;
+  riscv64) host_arch=riscv64 ;;
+  s390x) host_arch=s390x ;;
+  loongarch64 | mips | mips64 | ppc | ppc64 | sun4v) bail "$(uname -m) runner is not supported yet by this action; if you need support for this platform, please submit an issue at <https://github.com/taiki-e/install-action>" ;;
+  # TODO: uname -m on windows-11-arm returns "x86_64"
+  *) host_arch=x86_64 ;;
 esac
 
 tools=()